16-bits support

+++ src/sdl2/renderbase.hpp
@@ -26,9 +26,9 @@ protected:
 	SDL_Surface *surface;
 
     // Palette Lookup
-    uint32_t rgb[S16_PALETTE_ENTRIES * 3];    // Extended to hold shadow/hilight colours
+    uint16_t rgb[S16_PALETTE_ENTRIES * 3];    // Extended to hold shadow/hilight colours
 
-    uint32_t *screen_pixels;
+    uint16_t *screen_pixels;
 
     // Original Screen Width & Height
     uint16_t orig_width, orig_height;
